+++ pkg/anonymity/qb/anonymity_test.go
@@ -56,7 +56,9 @@ func TestNodeSettings(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	node, _, _ := testRunNodeWithDB(ctx, time.Minute, "", &tsDatabase{})
+	node, _, _ := testRunNodeWithDB(ctx, time.Minute, "", &tsDatabase{}, func(ctx context.Context, i INode, ik layer2.IParticipantKey, b []byte) ([]byte, error) {
+		return nil, nil
+	})
 	defer testFreeNodes([]INode{node}, 9)
 
 	sett := node.GetSettings()
@@ -72,7 +74,7 @@ func TestNodeSettings(t *testing.T) {
 			layer1.NewConstructSettings(&layer1.SConstructSettings{
 				FSettings: layer1.NewSettings(&layer1.SSettings{}),
 			}),
-			payload.NewPayload32(0, []byte{}),
+			[]byte{},
 		),
 	)
 	if err == nil {
@@ -107,7 +109,9 @@ func TestComplexFetchPayload(t *testing.T) {
 	defer cancel()
 
 	addresses := [2]string{testutils.TgAddrs[2], testutils.TgAddrs[3]}
-	nodes, privKeys := testRunNodes(ctx, t, time.Minute, addresses, 0)
+	nodes, privKeys := testRunNodes(ctx, t, time.Minute, addresses, 0, func(_ context.Context, _ INode, _ layer2.IParticipantKey, reqBytes []byte) ([]byte, error) {
+		return []byte(string(reqBytes) + " (response)"), nil
+	})
 	if nodes[0] == nil {
 		t.Fatal("nodes is null")
 	}
@@ -125,7 +129,7 @@ func TestComplexFetchPayload(t *testing.T) {
 			resp, err := nodes[0].FetchPayload(
 				ctx,
 				privKeys[1].GetPubKey(),
-				payload.NewPayload32(tcHead, []byte(reqBody)),
+				[]byte(reqBody),
 			)
 			if err != nil {
 				t.Errorf("%s (%d)", err.Error(), i)
@@ -150,7 +154,9 @@ func TestF2FWithoutFriends(t *testing.T) {
 
 	// 3 seconds for wait
 	addresses := [2]string{testutils.TgAddrs[10], testutils.TgAddrs[11]}
-	nodes, privKeys := testRunNodes(ctx, t, 3*time.Second, addresses, 1)
+	nodes, privKeys := testRunNodes(ctx, t, 3*time.Second, addresses, 1, func(_ context.Context, _ INode, _ layer2.IParticipantKey, reqBytes []byte) ([]byte, error) {
+		return []byte(string(reqBytes) + " (response)"), nil
+	})
 	if nodes[0] == nil {
 		t.Fatal("nodes is null")
 	}
@@ -163,7 +169,7 @@ func TestF2FWithoutFriends(t *testing.T) {
 	_, err := nodes[0].FetchPayload(
 		ctx,
 		privKeys[1].GetPubKey(),
-		payload.NewPayload32(tcHead, []byte(tcMsgBody)),
+		[]byte(tcMsgBody),
 	)
 	if err != nil {
 		return
@@ -179,24 +185,19 @@ func TestFetchPayload(t *testing.T) {
 	defer cancel()
 
 	addresses := [2]string{testutils.TgAddrs[12], testutils.TgAddrs[13]}
-	nodes, privKeys := testRunNodes(ctx, t, time.Minute, addresses, 4)
+	nodes, privKeys := testRunNodes(ctx, t, time.Minute, addresses, 4, func(_ context.Context, _ INode, _ layer2.IParticipantKey, reqBytes []byte) ([]byte, error) {
+		return []byte(fmt.Sprintf("echo: '%s'", string(reqBytes))), nil
+	})
 	if nodes[0] == nil {
 		t.Fatal("nodes is null")
 	}
 	defer testFreeNodes(nodes[:], 4)
 
-	nodes[1].HandleFunc(
-		tcHead,
-		func(_ context.Context, _ INode, _ layer2.IParticipantKey, reqBytes []byte) ([]byte, error) {
-			return []byte(fmt.Sprintf("echo: '%s'", string(reqBytes))), nil
-		},
-	)
-
 	largeBodySize := nodes[0].GetQBProcessor().GetScheme().GetPayloadLimit() - encoding.CSizeUint64 + 1
 	_, err := nodes[0].FetchPayload(
 		ctx,
 		privKeys[1].GetPubKey(),
-		payload.NewPayload32(tcHead, random.NewRandom().GetBytes(largeBodySize)),
+		random.NewRandom().GetBytes(largeBodySize),
 	)
 	if err == nil {
 		t.Fatal("success fetch payload with large body")
@@ -205,7 +206,7 @@ func TestFetchPayload(t *testing.T) {
 	result, err1 := nodes[0].FetchPayload(
 		ctx,
 		privKeys[1].GetPubKey(),
-		payload.NewPayload32(tcHead, []byte(tcMsgBody)),
+		[]byte(tcMsgBody),
 	)
 	if err1 != nil {
 		t.Fatal(err1)
@@ -222,28 +223,23 @@ func TestBroadcastPayload(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
+	chResult := make(chan string)
 	addresses := [2]string{testutils.TgAddrs[14], testutils.TgAddrs[15]}
-	nodes, privKeys := testRunNodes(ctx, t, time.Minute, addresses, 3)
+	nodes, privKeys := testRunNodes(ctx, t, time.Minute, addresses, 3, func(_ context.Context, _ INode, _ layer2.IParticipantKey, reqBytes []byte) ([]byte, error) {
+		res := fmt.Sprintf("echo: '%s'", string(reqBytes))
+		go func() { chResult <- res }()
+		return nil, nil
+	})
 	if nodes[0] == nil {
 		t.Fatal("nodes is null")
 	}
 	defer testFreeNodes(nodes[:], 3)
 
-	chResult := make(chan string)
-	nodes[1].HandleFunc(
-		tcHead,
-		func(_ context.Context, _ INode, _ layer2.IParticipantKey, reqBytes []byte) ([]byte, error) {
-			res := fmt.Sprintf("echo: '%s'", string(reqBytes))
-			go func() { chResult <- res }()
-			return nil, nil
-		},
-	)
-
 	largeBodySize := nodes[0].GetQBProcessor().GetScheme().GetPayloadLimit() - encoding.CSizeUint64 + 1
 	err := nodes[0].SendPayload(
 		context.Background(),
 		privKeys[1].GetPubKey(),
-		payload.NewPayload64(uint64(tcHead), random.NewRandom().GetBytes(largeBodySize)),
+		random.NewRandom().GetBytes(largeBodySize),
 	)
 	if err == nil {
 		t.Fatal("success broadcast payload with large body")
@@ -252,7 +248,7 @@ func TestBroadcastPayload(t *testing.T) {
 	err1 := nodes[0].SendPayload(
 		context.Background(),
 		privKeys[1].GetPubKey(),
-		payload.NewPayload64(uint64(tcHead), []byte(tcMsgBody)),
+		[]byte(tcMsgBody),
 	)
 	if err1 != nil {
 		t.Fatal(err1)
@@ -276,7 +272,9 @@ func TestEnqueuePayload(t *testing.T) {
 	defer cancel()
 
 	addresses := [2]string{testutils.TgAddrs[16], testutils.TgAddrs[17]}
-	nodes, privKeys := testRunNodes(ctx, t, time.Minute, addresses, 8)
+	nodes, privKeys := testRunNodes(ctx, t, time.Minute, addresses, 8, func(_ context.Context, _ INode, _ layer2.IParticipantKey, reqBytes []byte) ([]byte, error) {
+		return []byte(string(reqBytes) + " (response)"), nil
+	})
 	if nodes[0] == nil {
 		t.Fatal("nodes is null")
 	}
@@ -295,7 +293,7 @@ func TestEnqueuePayload(t *testing.T) {
 	}
 
 	pldBytes := payload.NewPayload64(
-		joinHead(sAction(1).setType(true), tcHead).uint64(),
+		setRequestBit(tcHead),
 		[]byte(tcMsgBody),
 	).ToBytes()
 
@@ -321,7 +319,9 @@ func TestHandleWrapper(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	_node, _, privKey := testRunNode(ctx, time.Minute, "", 7, 0)
+	_node, _, privKey := testRunNode(ctx, time.Minute, "", 7, 0, func(_ context.Context, _ INode, _ layer2.IParticipantKey, _ []byte) ([]byte, error) {
+		return nil, errors.New("some error") //nolint:err113
+	})
 	defer testFreeNodes([]INode{_node}, 7)
 
 	node := _node.(*sNode)
@@ -340,7 +340,7 @@ func TestHandleWrapper(t *testing.T) {
 	msg, err := scheme.EncryptMessage(
 		pubKey,
 		payload.NewPayload64(
-			joinHead(sAction(1).setType(true), tcHead).uint64(),
+			setRequestBit(tcHead),
 			[]byte(tcMsgBody),
 		).ToBytes(),
 	)
@@ -353,10 +353,6 @@ func TestHandleWrapper(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	netMsgX := node.testNewNetworkMessageWithInvalidNetworkMask(sett, msg)
-	if err := handler(ctx, netMsgX); err == nil {
-		t.Fatal("success handle message with invalid network mask")
-	}
 	netMsgY := node.testNewNetworkMessageWithAnotherMessageType(sett, msg)
 	if err := handler(ctx, netMsgY); err == nil {
 		t.Fatal("success handle message with another network type")
@@ -376,17 +372,10 @@ func TestHandleWrapper(t *testing.T) {
 		t.Fatal(err) // works only logger
 	}
 
-	node.HandleFunc(
-		111,
-		func(_ context.Context, _ INode, _ layer2.IParticipantKey, _ []byte) ([]byte, error) {
-			return nil, errors.New("some error") //nolint:err113
-		},
-	)
-
 	msg2, err := scheme.EncryptMessage(
 		pubKey,
 		payload.NewPayload64(
-			joinHead(sAction(1).setType(true), 111).uint64(),
+			setRequestBit(111),
 			[]byte(tcMsgBody),
 		).ToBytes(),
 	)
@@ -418,7 +407,7 @@ func TestHandleWrapper(t *testing.T) {
 	msg4, err := scheme.EncryptMessage(
 		pubKey,
 		payload.NewPayload64(
-			joinHead(sAction(1).setType(false), 111).uint64(),
+			setResponseBit(111),
 			[]byte(tcMsgBody),
 		).ToBytes(),
 	)
@@ -449,7 +438,9 @@ func TestStoreHashWithBroadcastMessage(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	_node, _, privKey := testRunNode(ctx, time.Minute, "", 6, 0)
+	_node, _, privKey := testRunNode(ctx, time.Minute, "", 6, 0, func(ctx context.Context, i INode, ik layer2.IParticipantKey, b []byte) ([]byte, error) {
+		return nil, nil
+	})
 	defer testFreeNodes([]INode{_node}, 6)
 
 	node := _node.(*sNode)
@@ -458,7 +449,7 @@ func TestStoreHashWithBroadcastMessage(t *testing.T) {
 	msg, err := scheme.EncryptMessage(
 		privKey.GetPubKey(),
 		payload.NewPayload64(
-			joinHead(sAction(1).setType(true), 111).uint64(),
+			setRequestBit(111),
 			[]byte(tcMsgBody),
 		).ToBytes(),
 	)
@@ -491,7 +482,9 @@ func TestRecvSendMessage(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	_node, _, privKey := testRunNode(ctx, time.Minute, "", 5, 0)
+	_node, _, privKey := testRunNode(ctx, time.Minute, "", 5, 0, func(ctx context.Context, i INode, ik layer2.IParticipantKey, b []byte) ([]byte, error) {
+		return nil, nil
+	})
 	defer testFreeNodes([]INode{_node}, 5)
 
 	node := _node.(*sNode)
@@ -500,7 +493,7 @@ func TestRecvSendMessage(t *testing.T) {
 	}
 
 	pubKey := privKey.GetPubKey()
-	actionKey := newActionKey(pubKey, sAction(111).setType(true))
+	actionKey := newActionKey(pubKey, setRequestBit(111))
 
 	node.setAction(actionKey)
 	action, ok := node.getAction(actionKey)
@@ -525,7 +518,7 @@ func TestRecvSendMessage(t *testing.T) {
 
 	msgBody := "hello, world!"
 	pldBytes := payload.NewPayload64(
-		joinHead(sAction(1).setType(true), tcHead).uint64(),
+		setRequestBit(tcHead),
 		[]byte(msgBody),
 	).ToBytes()
 
@@ -553,14 +546,14 @@ func TestRecvSendMessage(t *testing.T) {
 // nodes[2], nodes[3], nodes[4] = routes
 // nodes[2], nodes[4] are have open ports
 // Scheme: (nodes[0]) -> nodes[2] -> nodes[3] -> nodes[4] -> (nodes[1])
-func testRunNodes(ctx context.Context, t *testing.T, timeWait time.Duration, addresses [2]string, typeDB int) ([5]INode, [5]asymmetric.IPrivKey) {
+func testRunNodes(ctx context.Context, t *testing.T, timeWait time.Duration, addresses [2]string, typeDB int, handlerF IHandlerF) ([5]INode, [5]asymmetric.IPrivKey) {
 	nodes := [5]INode{}
 	networkNodes := [5]network.INode{}
 	privKeys := [5]asymmetric.IPrivKey{}
 	addrs := [5]string{"", "", addresses[0], "", addresses[1]}
 
 	for i := 0; i < 5; i++ {
-		nodes[i], networkNodes[i], privKeys[i] = testRunNode(ctx, timeWait, addrs[i], typeDB, i)
+		nodes[i], networkNodes[i], privKeys[i] = testRunNode(ctx, timeWait, addrs[i], typeDB, i, handlerF)
 		if nodes[i] == nil {
 			t.Errorf("node (%d) is not running %d", i, typeDB)
 			return [5]INode{}, [5]asymmetric.IPrivKey{}
@@ -572,16 +565,6 @@ func testRunNodes(ctx context.Context, t *testing.T, timeWait time.Duration, add
 
 	nodes[0].GetKeysContainer().Add(pubKey1)
 	nodes[1].GetKeysContainer().Add(pubKey0)
-
-	for _, node := range nodes {
-		node.HandleFunc(
-			tcHead,
-			func(_ context.Context, _ INode, _ layer2.IParticipantKey, reqBytes []byte) ([]byte, error) {
-				// send response
-				return []byte(string(reqBytes) + " (response)"), nil
-			},
-		)
-	}
 
 	go func() {
 		if err := networkNodes[2].Run(ctx); err != nil && !errors.Is(err, net.ErrClosed) && !errors.Is(err, context.Canceled) {
@@ -649,11 +632,10 @@ func (p *stLogging) HasErro() bool {
 }
 */
 
-func testRunNodeWithDB(ctx context.Context, timeWait time.Duration, addr string, db database.IKVDatabase) (INode, network.INode, asymmetric.IPrivKey) {
+func testRunNodeWithDB(ctx context.Context, timeWait time.Duration, addr string, db database.IKVDatabase, handlerF IHandlerF) (INode, network.INode, asymmetric.IPrivKey) {
 	privKey := asymmetric.NewPrivKey()
 	msgChan := make(chan layer1.IMessage)
 	parallel := uint64(1)
-	networkMask := uint32(1)
 	limitVoidSize := uint64(10_000)
 	networkNode := network.NewNode(
 		network.NewSettings(&network.SSettings{
@@ -672,16 +654,18 @@ func testRunNodeWithDB(ctx context.Context, timeWait time.Duration, addr string,
 				FWriteTimeout:          time.Minute,
 			}),
 		}),
+		func(_ context.Context, _ network.INode, _ conn.IConn, msg layer1.IMessage) error {
+			msgChan <- msg
+			return nil
+		},
 		cache.NewLRUCache(1024),
-	).HandleFunc(networkMask, func(_ context.Context, _ network.INode, _ conn.IConn, msg layer1.IMessage) error {
-		msgChan <- msg
-		return nil
-	})
+	)
 	node := NewNode(
 		NewSettings(&SSettings{
 			FServiceName:  "TEST",
 			FFetchTimeout: timeWait,
 		}),
+		handlerF,
 		// internal_std_logger.NewStdLogger(&stLogging{}, internal_anon_logger.GetLogFunc()),
 		logger.NewLogger(
 			logger.NewSettings(&logger.SSettings{}),
@@ -711,7 +695,6 @@ func testRunNodeWithDB(ctx context.Context, timeWait time.Duration, addr string,
 					}),
 					FParallel: parallel,
 				}),
-				FNetworkMask:  networkMask,
 				FQueuePoolCap: [2]uint64{tcQueueCap, tcQueueCap},
 				FQueuePeriod:  time.Second,
 				FConsumersCap: 1,
@@ -726,12 +709,12 @@ func testRunNodeWithDB(ctx context.Context, timeWait time.Duration, addr string,
 	return node, networkNode, privKey
 }
 
-func testRunNode(ctx context.Context, timeWait time.Duration, addr string, typeDB, numDB int) (INode, network.INode, asymmetric.IPrivKey) {
+func testRunNode(ctx context.Context, timeWait time.Duration, addr string, typeDB, numDB int, handlerF IHandlerF) (INode, network.INode, asymmetric.IPrivKey) {
 	db, err := database.NewKVDatabase(fmt.Sprintf(tcPathDBTemplate, typeDB, numDB))
 	if err != nil {
 		panic(err)
 	}
-	return testRunNodeWithDB(ctx, timeWait, addr, db)
+	return testRunNodeWithDB(ctx, timeWait, addr, db, handlerF)
 }
 
 func testFreeNodes(nodes []INode, typeDB int) {
@@ -750,20 +733,7 @@ func testDeleteDB(typeDB int) {
 func (p *sNode) testNewNetworkMessage(pSett layer1.IConstructSettings, pMsgBytes []byte) layer1.IMessage {
 	return layer1.NewMessage(
 		pSett,
-		payload.NewPayload32(
-			p.fQBProcessor.GetSettings().GetNetworkMask(),
-			pMsgBytes,
-		),
-	)
-}
-
-func (p *sNode) testNewNetworkMessageWithInvalidNetworkMask(pSett layer1.IConstructSettings, pMsgBytes []byte) layer1.IMessage {
-	return layer1.NewMessage(
-		pSett,
-		payload.NewPayload32(
-			p.fQBProcessor.GetSettings().GetNetworkMask()^1,
-			pMsgBytes,
-		),
+		pMsgBytes,
 	)
 }
 
@@ -774,10 +744,7 @@ func (p *sNode) testNewNetworkMessageWithAnotherMessageType(pSett layer1.IConstr
 				FNetworkKey: pSett.GetSettings().GetNetworkKey() + "_",
 			}),
 		}),
-		payload.NewPayload32(
-			p.fQBProcessor.GetSettings().GetNetworkMask()^1,
-			pMsgBytes,
-		),
+		pMsgBytes,
 	)
 }