+++ pkg/anonymity/qb/examples/echo/main.go
@@ -4,10 +4,6 @@ import (
 	"context"
 	"fmt"
 	"time"
-
-	anonymity "github.com/number571/go-peer/pkg/anonymity/qb"
-	"github.com/number571/go-peer/pkg/crypto/scheme/layer2"
-	"github.com/number571/go-peer/pkg/payload"
 )
 
 const (
@@ -30,7 +26,7 @@ func main() {
 		resp, _ := nodeClient.fAnonymity.FetchPayload(
 			ctx,
 			keyToService,
-			payload.NewPayload32(nodeRouter, []byte("hello, world!")),
+			[]byte("hello, world!"),
 		)
 		fmt.Println(string(resp))
 	}
@@ -49,12 +45,6 @@ func runClientNode() *sNode {
 func runServiceNode() *sNode {
 	ctx := context.Background()
 	node := newNode("snode", nodeAddress)
-	node.fAnonymity.HandleFunc(
-		nodeRouter,
-		func(_ context.Context, _ anonymity.INode, _ layer2.IParticipantKey, b []byte) ([]byte, error) {
-			return []byte("echo: " + string(b)), nil
-		},
-	)
 
 	go func() { _ = node.fAnonymity.Run(ctx) }()
 	go func() { _ = node.fNetwork.Run(ctx) }()