+++ pkg/anonymity/qb/examples/ping-pong/main.go
@@ -2,40 +2,13 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"time"
 
-	anonymity "github.com/number571/go-peer/pkg/anonymity/qb"
-	"github.com/number571/go-peer/pkg/crypto/scheme/layer2"
 	"github.com/number571/go-peer/pkg/encoding"
-	"github.com/number571/go-peer/pkg/payload"
 )
 
 const (
 	nodeAddress = "127.0.0.1:8080"
-	nodeRouter  = uint32(0xA557711A)
-)
-
-var (
-	handler = func(ctx context.Context, n anonymity.INode, pKey layer2.IParticipantKey, b []byte) ([]byte, error) {
-		numBytes := [encoding.CSizeUint64]byte{}
-		copy(numBytes[:], b)
-
-		num := encoding.BytesToUint64(numBytes)
-		msg := "ping"
-		if num%2 == 1 {
-			msg = "pong"
-		}
-		fmt.Printf("%s-%d\n", msg, num)
-
-		numBytes = encoding.Uint64ToBytes(num + 1)
-		_ = n.SendPayload(
-			ctx,
-			pKey,
-			payload.NewPayload64(uint64(nodeRouter), numBytes[:]),
-		)
-		return nil, nil
-	}
 )
 
 func init() {
@@ -52,7 +25,7 @@ func main() {
 	_ = nodeClient.fAnonymity.SendPayload(
 		context.Background(),
 		keyService,
-		payload.NewPayload64(uint64(nodeRouter), numBytes[:]),
+		numBytes[:],
 	)
 
 	select {}
@@ -61,7 +34,6 @@ func main() {
 func runClientNode() *sNode {
 	ctx := context.Background()
 	node := newNode("cnode", "")
-	node.fAnonymity.HandleFunc(nodeRouter, handler)
 
 	go func() { _ = node.fAnonymity.Run(ctx) }()
 	_ = node.fNetwork.AddConnection(ctx, nodeAddress)
@@ -72,7 +44,6 @@ func runClientNode() *sNode {
 func runServiceNode() *sNode {
 	ctx := context.Background()
 	node := newNode("snode", nodeAddress)
-	node.fAnonymity.HandleFunc(nodeRouter, handler)
 
 	go func() { _ = node.fAnonymity.Run(ctx) }()
 	go func() { _ = node.fNetwork.Run(ctx) }()