+++ pkg/anonymity/qb/types.go
@@ -8,7 +8,6 @@ import (
 	"github.com/number571/go-peer/pkg/anonymity/qb/queue"
 	"github.com/number571/go-peer/pkg/crypto/scheme/layer2"
 	"github.com/number571/go-peer/pkg/logger"
-	"github.com/number571/go-peer/pkg/payload"
 	"github.com/number571/go-peer/pkg/storage/database"
 	"github.com/number571/go-peer/pkg/types"
 )
@@ -19,7 +18,6 @@ type (
 
 type INode interface {
 	types.IRunner
-	HandleFunc(uint32, IHandlerF) INode
 
 	GetLogger() logger.ILogger
 	GetSettings() ISettings
@@ -28,8 +26,8 @@ type INode interface {
 	GetKeysContainer() layer2.IKeysContainer
 	GetQBProcessor() queue.IQBProblemProcessor
 
-	SendPayload(context.Context, layer2.IParticipantKey, payload.IPayload64) error
-	FetchPayload(context.Context, layer2.IParticipantKey, payload.IPayload32) ([]byte, error)
+	SendPayload(context.Context, layer2.IParticipantKey, []byte) error
+	FetchPayload(context.Context, layer2.IParticipantKey, []byte) ([]byte, error)
 }
 
 type ISettings interface {